Adding sort section to config file.
Default sorted by threads... I would prefer to sort by reverse-date of
last message in thread (you know, like G-mail), but that doesn't seem
possible. The "next best thing" according to the Interwebz is commented
out for now. It seems like it might be easier to just /~N or * and scroll
up.