Fix the ssh_agent_setup function.
I had the if statement list b0rked.

if [ ! -f "$ssh_agent_env" -o \
       ! ps ... ]; then ...

The [ or test command is for testing argument or file system conditions.
It does not execute the parts of its conditions as commands. >:( If you
want to execute a process, like ps, to check its success or failure then
you do so directly, without [. So I basically just had to replace the
above with:

if [ ! -f "$ssh_agent_env" ] ||
        ! ps ...; then ...

For some reason I could swear I already had this corrected for me a while
back, but perhaps it was on different code or perhaps I was just
intoxicated and forgot to fix it. :P I digress... It is fixed now, and
ssh_agent_setup seems to work as intended now!