Add server.validate = true to hgrc.
It seems Mercurial doesn't verify the validity of changesets by default.
This lead to corruption of two repositories at j0rb, and that's more than
I'm comfortable with. :) Hopefully it doesn't affect performance
drastically, but I think the peace of knowing that the software is at
least checking that things look OK is probably worth the hit..

(I think Git does this by default and it's fast... Pffft..)