Add log.. alias to hgrc.
The intent is to achieve something like git log branch1..branch2, though
it's not really the same thing since Mercurial's backend works
differently. It tries to find the earliest common ancestor between the
specified *commits*, and then get the logs for the commits on branch2
that have happened since. I'm not sure it's correct, but it seemed to
work in my immediate use case. I don't do a lot of branch/merging in
Mercurial (for obvious reasons, I think), but hopefully this will make
it a bit easier when I do.