Fixing basic logic of killall bash function.
I don't even remember when I wrote this, but it was probably 2 or 3
years ago... Evidently I forgot all about it because it doesn't work. :P
I am amazed that it hasn't interfered with things in all of this time.
I'm sure that I have been using killall and it should have been loaded
in my environment... Maybe I thought I was killing things, but wasn't.
Hell, maybe it failed and I was blaming the system command instead of my
bash function. It's funny if you think about it... I'm not really sure
what this function is even for since it seems to more or less reinvent
the square wheel that is killall, but I assume that I intended to make
it more user friendly (i.e., not loop indefinitely until you ^C).

I could probably enhance this quite a bit with my current level of
experience with bash (and Perl, for that matter), but perhaps I should
rename it so there is never a conflict with the alternative command...