Add a bash function to xinit steam.
Steam seems to behave somewhat weird in xmonad, plus I'm still debugging
how to get full-screen games to pick a usable resolution. Means a lot of
experimentation. I can't even get full-screen games to show up in an
xmonad session. They just seem to hang on a black screen. From an X
server running just steam, however, the games do run, but at an
extremely small resolution which is completely unsable for several
reasons. This is just a helper to debug and get it working... Might
eventually not be necessary.