Remove space from VBScript SQL helper command.
I'm not sure if this is desirable or not, but it is what I have been
using at j0rb (which is the only time I edit VBScript code..). The
indentation should rightly be 4 spaces, but 3 spaces seems to work or
align better or something... I don't know... I think that 3 spaces is
what you get if you just hit tab due to the fixed depth of indentation
in Vim, so using 3 spaces for these commands means that I get the same
result whether I type by hand or use these shortcuts.. Or something.
h4x.