h4x: Enhanced PS1 using git and music-control.
The only way to do it without screwing up the prompt length count is
with PROMPT_COMMAND. The quoting/escaping can get quite ugly.
Unfortunately, it seems the most readable solution is a single
single-quoted literal string. This then holds the command to execute,
that which is wrapped in a single double-quoted string. The usual rules
there apply. All special characteres meant to be interpretted literally
need to be escaped (include backslashes themselves). Finally seem to
have gotten it work... Unfortunately it's at 158 columns, and there's no
pretty way to wrap it...