Add mailbox for ingydotnet/git-hub.
As I recall, that is a custom subcommand for git to communicate with
git-hub. I like the idea of eliminating use of the Web site and managing
everything through a CLI instead of a Web browser, but for now I haven't
really even used it. It's good to see that it is active at least.