Add grep-based aliase and functions.
I probably should be using ack most of the time. A not too distance goal
will be to port all shell scripts to use ack so that I can enable the
commented out function to blatantly fail when grep is used. Though that
might still be a bad idea for building third party software and the
like...