Configure Mutt to store sent messages in inbox.
At least, that's what I hope this will do. :) I like for threads to stay
together and it's easier for threads that involve replies that end up in
my inbox to do so if my sent messages also are written to my inbox. I
got used to this from Web mail, and particularly Google mail.