Extend wifi function to change networks.
This is rudimentary and manual, but it suits me fine on my netbook. I
have to su to root to manage my networks in /etc/network/interfaces.d/,
and then I need to visudo sudoers in order to allow myself to symlink
/etc/network/interfaces.d/wlan0, but then I can relatively easy switch
between networks as I move around. It doesn't happen often enough to be
annoying, and it keeps me in control.