h4x: Add a bash watchdog function.
I used this for deluge since it seemed insistant on crashing regularly.
I can't figure out what is causing it to crash yet, but it may be a bug
in libtorrent or something like that. In any case, the watchdog helps to
notice its death and replace it with a new instance.

It's a 1 minute hack so no guarantees on reliability or
user-friendliness.