h4x: Commit irssi config changes.
My irssi config appeared to corrupt itself because irssi hung and when I
relaunched it it started to segfault. I saved that for later, maybe
interesting to dig into it, but I've restore the original config. Here
are a few key bindings which never got tracked... Used to clear the
buffer of noise. I don't find the smartfilter script does a very good
job, but I haven't made any attempts to configure it or fix it...