h4x: Make Mercurial log branch names yellow?
Mercurial config changes are almost always coming from Windows at j0rb
and so there's no guarantee that they'll actually behave properly in a
sane environment. They are more or less hacked into the .hgrc, and due
to symlink misfeatures in Windows they are typically not immediately
recognized in the working tree of the repo and so they pile on before I
get a chance to sort though them and add them... Many of these changes
were even copied from a different workstation.

What I'm getting at is that I'm not sure if this color change makes
sense without other things (templates, or something, for example).