Replace needless ack in ssh-agent-setup with grep.
I probably changed this when I began my switch to ack for everyday usage
since for interactive operations it's generally much more friendly. That
said, for automated scripts such as this one, and particularly the
receiving end of a pipe, the advantages of ack aren't really there and
grep has its own advantages: it's universal and more efficient.