Inject ssh-add-if-empty into social tmux session.
Since several of the panes in the social session attempt to SSH it makes
sense to have the ssh-agent setup *before* we start firing off those
commands. This way we can try to avoid having to enter the credentials
several times or interfering with the initialization of the panes.

This may need some work in graphical environments that use a different
agent... I'll need to experiment with that when I'm in such an
environment next.