h4x: Create function to define more 'py' sessions.
These are essentially tmux sessions with two vertical panes, on one side
a python REPL and the other a shell. They are essentially just scratch
environments, but the python interpreter is intended to remain open and
available for whatever you happen to want it for (math, hacking,
experimenting, etc.). I aim to learn more Python and especially Python
modules/libraries/frameworks so this is intended to help with that. This
particular function was essentially a contrived idea where I already had
a long-running command going in the shell of my py session and it
occurred to me that I could create a new one... Note that the alias only
exists within the current shell currently. Nothing will be done to
spread them to new or existing shells... To access them there you'd need
to repeat the mkpy command in each such shell...