h4x: Add a lisp repl loop function.
I haven't seen this in months, but it appears the idea was a persistent
code file that would be repeatedly run in an interactive REPL. You loop
back and forth between the editor. And the code file is persisted to
disk. Probably was a hack to solve the problem of experimenting in the
REPL, but then having to re-type the code into a code file after...
There must be better solutions out there.