install: Comment out the evolve repo.
I haven't been working with Mercurial for a while (sadly), but it seems
that evolve is no longer hosted at this address. I suspect it might have
been partially included with the official distro, but whether or not
that's true it is at least not at this address. When I next need it I'll
have to figure out how to get it (actually, if memory serves, I think
it's a pip package now).